composition_ace/pyproject.toml
[project]
name = "composition-ace"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai==1.96.1",
    "pydantic>=2.12.3",
    "python-dotenv>=1.2.1",
]
